[RAM-4] Refactor move logic data back into data directory
