+++ Assets/Source/Logic/Data/BattlefieldResource.cs.meta
