+++ Assets/Source/Logic/Data/Player.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Source.Logic
+namespace Source.Logic.Data
 {
     [Serializable]
     public class Player
