+++ Assets/Source/Serialization/Data/BattlefieldItemData.cs
@@ -1,5 +1,6 @@
 using System;
 using Source.Logic;
+using Source.Logic.Data;
 
 namespace Source.Serialization.Data
 {
