+++ Assets/Source/Visuals/DataItemVisual.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Source.Interactions;
 using Source.Logic;
+using Source.Logic.Data;
 using UnityEngine;
 
 namespace Source.Visuals
