+++ Assets/Source/Visuals/PlayerItemStorageVisual.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Source.Input;
 using Source.Logic;
+using Source.Logic.Data;
 using Source.Serialization;
 using UnityEngine;
 using UnityEngine.UI;
